+++ src/app/learn/_learn/theory/interslavic/history-of-slavic-languages/history-of-slavic-languages.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { LearnPageTemplateComponent } from "../../../../learn-page-template/learn-page-template.component";
+import { LearnPageTemplateComponent } from "../../../../page-template/page-template.component";
 import { RouterLink } from "@angular/router";
 
 @Component({
