--- src/app/learn/learn-navigation-compact/learn-navigation-compact.component.html
@@ -1,44 +1,0 @@
-<div class="learn-navigation-compact">
-
-  <div class="current-structure">
-    <span id="learn-icon"
-      class="learn-navigation-compact-icon material-symbols-rounded"
-      routerLink="/learn"
-    >school</span>
-
-    <ng-container *ngFor="let elem of this.structure; index as i">
-      <span class="learn-navigation-compact-arrow material-symbols-rounded"
-        [ngClass]="i === this.structure.length - 1 ? 'active' : ''"
-      >chevron_right</span>
-
-      <button class="learn-navigation-compact-button" id="learn-navigation-compact-button-{{i}}"
-        (click)="this.revealOtherOptions(i)"
-      >
-        <span class="dropdown-title">{{ elem }}</span>
-        <span class="expand-arrow material-symbols-rounded"
-          *ngIf="i <= (this.structure.length - 1) && adjacent(i).all.length > 1"
-        >unfold_more</span>
-      </button>
-    </ng-container>
-  </div>
-
-  <ng-container *ngFor="let elem of this.structure; index as i">
-    <div class="other-options" id="{{this.id_prefix + i}}">
-      <ng-container *ngFor="let adj of this.adjacentValuesOfEachStructureElement[i].top">
-        <div class="top">
-          <a [routerLink]="(i === 1 ? ('../../' + this.learnRouting.makeURLSafe(this.structure[0])) : ('/learn/theory')) + '/' + this.learnRouting.makeURLSafe(adj)">{{adj}}</a>
-        </div>
-      </ng-container>
-      <ng-container *ngIf="this.adjacentValuesOfEachStructureElement[i]._self">
-        <div class="self">
-          <a>{{this.adjacentValuesOfEachStructureElement[i]._self}}</a>
-        </div>
-      </ng-container>
-      <ng-container *ngFor="let adj of this.adjacentValuesOfEachStructureElement[i].bottom" >
-        <div class="bottom">
-          <a [routerLink]="(i === 1 ? ('../../' + this.learnRouting.makeURLSafe(this.structure[0])) : ('/learn/theory')) + '/' + this.learnRouting.makeURLSafe(adj)">{{adj}}</a>
-        </div>
-      </ng-container>
-    </div>
-  </ng-container>
-</div>
