--- src/app/learn/learn-navigation-compact/learn-navigation-compact.component.scss
@@ -1,73 +1,0 @@
-.learn-navigation-compact {
-  display: flex;
-  flex-flow: column nowrap;
-}
-
-.other-options {
-  display: none;
-  &.show {
-    display: flex;
-  }
-  gap: 0.5rem;
-  flex-flow: column nowrap;
-  background-color: var(--background);
-  padding: 1.5rem 2rem 2rem 2rem;
-  a {
-    text-decoration: solid underline;
-  }
-  .self a {
-    color: var(--text) !important;
-  }
-  border-bottom: 0.125rem solid var(--border);
-}
-
-.current-structure {
-  display: flex;
-  flex-flow: row wrap;
-  align-items: center;
-
-  > :last-child {
-    background-color: var(--border) !important;
-  }
-
-  .learn-navigation-compact-icon {
-    padding: 0.5rem;
-  }
-  .learn-navigation-compact-button {
-    .dropdown-title {
-      font-weight: 600;
-    }
-
-    padding: 0.6rem;
-    border-radius: 1rem;
-
-    display: flex;
-    flex-flow: row nowrap;
-    align-items: center;
-    justify-content: center;
-    gap: 0.25rem;
-
-    &.show {
-      background-color: var(--secondary) !important;
-      color: var(--background) !important;
-    }
-
-    &, &:hover {
-      background-color: #00000000;
-      color: var(--text);
-    }
-  }
-
-  .learn-navigation-compact-icon {
-    margin:  0.5rem;
-    border-radius: 50%;
-    background-color: var(--accent);
-    color: var(--background);
-    @media (hover: hover) and (pointer: fine) {
-      &:hover {
-        cursor: pointer;
-      }
-    }
-  }
-}
-
