--- src/app/learn/learn-navigation-side/learn-navigation-side.component.html
@@ -1,26 +1,0 @@
-<div class="learn-navigation-side">
-  <div class="learn-navigation-side-elem top-level"
-    *ngFor="let parentElem of Object.keys(this.learnRouting.TheoryStructure)"
-    #parent (click)="revealChildren(parent)"
-  >
-    <div class="content">
-      <div class="learn-navigation-side-title">
-        <span class="material-symbols-rounded icon">chevron_right</span>
-        {{ parentElem }}
-      </div>
-    </div>
-    <div class="children-expand-animation-container">
-      <div class="children">
-        <div class="child"
-          *ngFor="let childElem of Object.keys(this.learnRouting.TheoryStructure[parentElem])"
-          [routerLink]="'../../' + this.learnRouting.makeURLSafe(parentElem) + '/' + this.learnRouting.makeURLSafe(childElem)"
-        >
-          <div class="content">
-            <div class="learn-navigation-side-title"><a>{{ childElem }}</a></div>
-          </div>
-        </div>
-      </div>
-    </div>
-  </div>
-</div>
-
