+++ src/app/learn/page-template/page-template.component.ts
@@ -1,14 +1,14 @@
 import { Component } from '@angular/core';
 import { SideProgressBarComponent } from "../side-progress-bar/side-progress-bar.component";
-import { LearnNavigationCompactComponent } from "../learn-navigation-compact/learn-navigation-compact.component";
-import { LearnNavigationSideComponent } from "../learn-navigation-side/learn-navigation-side.component";
+import { LearnNavigationCompactComponent } from "../navigation-compact/navigation-compact.component";
+import { LearnNavigationSideComponent } from "../navigation-side/navigation-side.component";
 
 @Component({
   selector: 'learn-page-template',
   standalone: true,
   imports: [ SideProgressBarComponent, LearnNavigationCompactComponent, LearnNavigationSideComponent ],
-  templateUrl: './learn-page-template.component.html',
-  styleUrl: './learn-page-template.component.scss',
+  templateUrl: './page-template.component.html',
+  styleUrl: './page-template.component.scss',
 })
 export class LearnPageTemplateComponent {
   getAllPageIds(level: number): string[] {
